re-organized and added more into DeviceStore
