+++ wolfie_home/static/js/views/devices/.babelrc
@@ -1,0 +1,1 @@
+{ "presets": ["react"] }
