+++ wolfie_home/static/js/views/devices/DeviceView.jsx
@@ -1,0 +1,1 @@
+DeviceView.jsx
