+++ wolfie_home/static/js/views/devices/build.sh
@@ -1,0 +1,3 @@
+#! /usr/bin/env bash
+
+/home/chaojie/Repo/ProtoPad/node_modules/.bin/babel DeviceListView.jsx > devices.js
