+++ wolfie_home/static/js/views/house/.babelrc
@@ -1,0 +1,1 @@
+{ "presets": ["react"] }
