+++ wolfie_home/templates/default_layout.html
@@ -58,7 +58,7 @@
     </div>
 
   </body>
-  <script type="text/javascript" src="{{ url_for('static', filename='js/user.js') }}"></script>
+  <script type="text/javascript" src="{{ url_for('static', filename='js/others/user.js') }}"></script>
   {% block js_at_the_end %}
   {% endblock %}
 </html>
