+++ wolfie_home/templates/login.html
@@ -16,7 +16,7 @@
              src="{{ url_for('static', filename='js/libraries/material.js') }}"></script>
     <script type="text/javascript" 
              src="{{ url_for('static', filename='js/libraries/ripples.js') }}"></script>
-    <script type="text/javascript" src="{{ url_for('static', filename='js/user.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/others/user.js') }}"></script>
 </head>
 
 <body>
